task: separate docs menu into its own component
