--- src/shell/components/GlobalAccountMenu/util.ts
@@ -1,4 +1,0 @@
-export const handleOpenUrl = (url: string, closeMenuCb?: () => void) => {
-  closeMenuCb && closeMenuCb();
-  window.open(url, "_blank", "noopener");
-};
